added disable key & automatic detection of pdfpages
pdfpages version >= 2021/01/09 v0.5r sets pdfpages@includegraphics@status
to indicate that it wants to get the number of pages. In that case, we
fall back to native \includegraphics. Users of older pdfpages can use
the explicit disable key:

\pgfkeys{/graphicscache/disable=true}
\includepdf[pages=1-3,disable=false]{mypdf.pdf}
\pgfkeys{/graphicscache/disable=false}

Note that passing "disable=false" to \includepdf enables graphicscache
for the final PDF inclusion. That way, we do not lose compression
and caching.